chore(boards):refactor
